https://github.com/mP1/walkingkooka-spreadsheet/pull/8293 value package extracted
- https://github.com/mP1/walkingkooka-spreadsheet/pull/8293
- value package extracted
