+++ src/main/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionCellFormattedValue.java
@@ -17,7 +17,7 @@
 
 package walkingkooka.spreadsheet.expression.function;
 
-import walkingkooka.spreadsheet.SpreadsheetCell;
+import walkingkooka.spreadsheet.value.SpreadsheetCell;
 import walkingkooka.tree.text.TextNode;
 
 /**
