+++ src/main/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionCellStyle.java
@@ -17,7 +17,7 @@
 
 package walkingkooka.spreadsheet.expression.function;
 
-import walkingkooka.spreadsheet.SpreadsheetCell;
+import walkingkooka.spreadsheet.value.SpreadsheetCell;
 import walkingkooka.tree.text.TextStyle;
 
 /**
