+++ src/main/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionNumberType.java
@@ -18,10 +18,10 @@
 package walkingkooka.spreadsheet.expression.function;
 
 import walkingkooka.collect.set.Sets;
-import walkingkooka.spreadsheet.HasSpreadsheetErrorKind;
 import walkingkooka.spreadsheet.expression.SpreadsheetExpressionEvaluationContext;
 import walkingkooka.spreadsheet.reference.SpreadsheetCellRangeReference;
 import walkingkooka.spreadsheet.reference.SpreadsheetCellReference;
+import walkingkooka.spreadsheet.value.HasSpreadsheetErrorKind;
 import walkingkooka.tree.expression.ExpressionNumber;
 import walkingkooka.tree.expression.function.ExpressionFunctionParameter;
 import walkingkooka.tree.expression.function.ExpressionFunctionParameterKind;
