+++ src/main/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionObjectCellTypeInfo.java
@@ -18,9 +18,9 @@
 package walkingkooka.spreadsheet.expression.function;
 
 import walkingkooka.NeverError;
-import walkingkooka.spreadsheet.SpreadsheetCell;
 import walkingkooka.spreadsheet.expression.SpreadsheetExpressionEvaluationContext;
 import walkingkooka.spreadsheet.reference.SpreadsheetCellReference;
+import walkingkooka.spreadsheet.value.SpreadsheetCell;
 import walkingkooka.text.CharSequences;
 import walkingkooka.tree.text.TextAlign;
 import walkingkooka.tree.text.TextStylePropertyName;
