+++ src/test/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionCellDateTimeSymbolsTest.java
@@ -19,7 +19,7 @@ package walkingkooka.spreadsheet.expression.function;
 
 import org.junit.jupiter.api.Test;
 import walkingkooka.datetime.DateTimeSymbols;
-import walkingkooka.spreadsheet.SpreadsheetCell;
+import walkingkooka.spreadsheet.value.SpreadsheetCell;
 
 import java.text.DateFormatSymbols;
 import java.util.Locale;
