+++ src/test/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionNaTest.java
@@ -19,9 +19,9 @@ package walkingkooka.spreadsheet.expression.function;
 
 import org.junit.jupiter.api.Test;
 import walkingkooka.collect.list.Lists;
-import walkingkooka.spreadsheet.SpreadsheetError;
-import walkingkooka.spreadsheet.SpreadsheetErrorKind;
 import walkingkooka.spreadsheet.expression.SpreadsheetExpressionEvaluationContext;
+import walkingkooka.spreadsheet.value.SpreadsheetError;
+import walkingkooka.spreadsheet.value.SpreadsheetErrorKind;
 
 public final class SpreadsheetExpressionFunctionNaTest extends SpreadsheetExpressionFunctionTestCase<SpreadsheetExpressionFunctionNa, SpreadsheetError> {
