+++ src/test/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionNumberCountBlankTest.java
@@ -19,9 +19,9 @@ package walkingkooka.spreadsheet.expression.function;
 
 import org.junit.jupiter.api.Test;
 import walkingkooka.collect.list.Lists;
-import walkingkooka.spreadsheet.SpreadsheetError;
 import walkingkooka.spreadsheet.expression.SpreadsheetExpressionEvaluationContext;
 import walkingkooka.spreadsheet.reference.SpreadsheetSelection;
+import walkingkooka.spreadsheet.value.SpreadsheetError;
 
 import java.time.LocalDateTime;
