+++ src/test/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionNumberTypeTest.java
@@ -19,9 +19,9 @@ package walkingkooka.spreadsheet.expression.function;
 
 import org.junit.jupiter.api.Test;
 import walkingkooka.collect.list.Lists;
-import walkingkooka.spreadsheet.SpreadsheetErrorKind;
 import walkingkooka.spreadsheet.expression.SpreadsheetExpressionEvaluationContext;
 import walkingkooka.spreadsheet.reference.SpreadsheetSelection;
+import walkingkooka.spreadsheet.value.SpreadsheetErrorKind;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
