+++ src/test/java/walkingkooka/spreadsheet/expression/function/SpreadsheetExpressionFunctionObjectErrorTypeTest.java
@@ -19,7 +19,7 @@ package walkingkooka.spreadsheet.expression.function;
 
 import org.junit.jupiter.api.Test;
 import walkingkooka.collect.list.Lists;
-import walkingkooka.spreadsheet.SpreadsheetErrorKind;
+import walkingkooka.spreadsheet.value.SpreadsheetErrorKind;
 import walkingkooka.tree.expression.ExpressionNumberKind;
 
 public final class SpreadsheetExpressionFunctionObjectErrorTypeTest extends SpreadsheetExpressionFunctionObjectTestCase<SpreadsheetExpressionFunctionObjectErrorType> {
